auto_tracker_pro/requirements.txt
ttkbootstrap==1.10.1
Pillow==10.0.1
mysql-connector-python==8.1.0
Werkzeug==2.3.7
requests==2.31.0